requirements.txt
future==0.17.1
google-auth==1.17.2
google-cloud-bigquery==1.25.0
google-cloud-core==1.3.0
matplotlib==3.1.0
numpy==1.16.4
pandas==0.24.2
pandas-gbq==0.13.2
pyarrow==0.17.1
regex==2020.6.8
scikit-learn==0.21.2
snorkel==0.9.5